sceneListeClient

+++ src/Controlleur/EcouteurListeClients.java
@@ -5,7 +5,6 @@ import com.jfoenix.controls.*;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
@@ -17,17 +16,13 @@ import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
+import javafx.stage.FileChooser;
 import javafx.stage.Stage;
 import javafx.util.Duration;
 import org.controlsfx.control.Notifications;
-
-import javax.management.Notification;
-import java.io.IOException;
+import java.io.*;
 import java.net.URL;
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
+import java.sql.*;
 import java.util.ResourceBundle;
 
 public class EcouteurListeClients implements Initializable {
@@ -39,7 +34,6 @@ public class EcouteurListeClients implements Initializable {
     @FXML private TableColumn<Client, String> colPrenom;
     @FXML private TableColumn<Client, String> colMail;
     @FXML private TableColumn<Client, Boolean> colFidele;
-
     ObservableList<Client> obList= FXCollections.observableArrayList();
 
     /*Concerne la scene*/
@@ -47,6 +41,11 @@ public class EcouteurListeClients implements Initializable {
     @FXML private JFXTextField prenom;
     @FXML private JFXCheckBox fidele;
     @FXML private JFXTextField mail;
+    @FXML private ImageView image;
+    private FileInputStream fis=new FileInputStream(new File("src/Images/Photo-non-disponible.jpg"));
+    private Boolean uneImageEstSelectionner=false;
+    public EcouteurListeClients() throws FileNotFoundException {
+    }
 
     public void remplirLaListe() throws SQLException
     {
@@ -75,13 +74,18 @@ public class EcouteurListeClients implements Initializable {
         table.getItems().clear();
     }
     public void supprimerClient() throws SQLException {
-        Client clientSupprmier=table.getSelectionModel().getSelectedItem();
-        ConnectionClass connectionClass=new ConnectionClass();
-        Connection connection=connectionClass.getConnection();
-        String sql="DELETE FROM client where idClient='"+clientSupprmier.getIdClient()+"'";
-        Statement statement = connection.createStatement();
-        statement.execute(sql);
-        nettoyageScene();
+        if(!table.getSelectionModel().isEmpty()) {
+            Client clientSupprmier = table.getSelectionModel().getSelectedItem();
+            ConnectionClass connectionClass = new ConnectionClass();
+            Connection connection = connectionClass.getConnection();
+            String sql = "DELETE FROM client where idClient='" + clientSupprmier.getIdClient() + "'";
+            Statement statement = connection.createStatement();
+            statement.execute(sql);
+            notifBuilder("Opération réussie",
+                    "Votre opération de suppression du " + nom.getText() + " est éffectué avec succès.",
+                    "/Images/checked.png");
+            nettoyageScene();
+        }
     }
     public void ajoutClient() throws SQLException {
         ConnectionClass connectionClass=new ConnectionClass();
@@ -96,10 +100,23 @@ public class EcouteurListeClients implements Initializable {
         }
         else{
             int fid=fidele.isSelected()?1:0;
-            String insertReq="INSERT INTO client (nomClient,prenomClient,clientFidele,mailClient) values " +
-                    "( '" + nom.getText()+"' , '"+prenom.getText()+"' , "+fid+" , '"+mail.getText()+"' )";
-            Statement statInsert = connection.createStatement();
-            statInsert.executeUpdate(insertReq);
+            String insertReq="INSERT INTO client (nomClient,prenomClient,clientFidele,pdp,mailClient) values (?,?,?,?,?)";
+            PreparedStatement statInsert = connection.prepareStatement(insertReq);
+            statInsert.setString(1,nom.getText());
+            statInsert.setString(2,prenom.getText());
+            statInsert.setInt(3,fid);
+
+            if(uneImageEstSelectionner){
+                statInsert.setBinaryStream(4,fis);
+                uneImageEstSelectionner=false;
+            }
+            else{
+                InputStream is=inputClient(new Client(nom.getText(),prenom.getText(),mail.getText(),fidele.isSelected()));
+                statInsert.setBinaryStream(4,is);
+            }
+
+            statInsert.setString(5,mail.getText());
+            statInsert.executeUpdate();
             statInsert.close();
             notifBuilder("Opération réussie",
                     "Votre opération d'ajouter le client "+nom.getText()+" est éffectué avec succès.",
@@ -108,15 +125,35 @@ public class EcouteurListeClients implements Initializable {
         }
 
     }
+    public void ChoisirUneImage() throws FileNotFoundException {
+        FileChooser fc = new FileChooser();
+        FileChooser.ExtensionFilter ext1 = new FileChooser.ExtensionFilter("JPG files(*.jpg)","*.JPG");
+        FileChooser.ExtensionFilter ext2 = new FileChooser.ExtensionFilter("PNG files(*.png)","*.PNG");
+        fc.getExtensionFilters().addAll(ext1,ext2);
+        File fichierSelect=fc.showOpenDialog(null);
+        if(fichierSelect!=null){
+            fis=new FileInputStream(fichierSelect);
+            image.setImage(new Image(fichierSelect.toURI().toString()));
+            uneImageEstSelectionner=true;
+        }
 
-    public void modifierClient() throws SQLException {
+
+    }
+    public void viderImage() throws IOException {
+        File file=new File("src/Images/Photo-non-disponible.jpg");
+        fis=new FileInputStream(file);
+        image.setImage(new Image("/Images/Photo-non-disponible.jpg"));
+        uneImageEstSelectionner=true;
+    }
+
+    public void modifierClient() throws SQLException, IOException {
         if(!table.getSelectionModel().isEmpty()){
             Client client=table.getSelectionModel().getSelectedItem();
             Client clientSelectionner=new Client(client.getIdClient(),nom.getText(),prenom.getText(),mail.getText(),fidele.isSelected());
-            if(!client.equals(clientSelectionner)) {
+            if(!client.equals(clientSelectionner) || uneImageEstSelectionner) {
                modifierclientSelectionner(clientSelectionner);
                notifBuilder("Opération réussie",
-                       "Votre opération de modifier le client "+client.getNomClient()+" a réussie.",
+                       "Votre opération de modifier le client "+client.getNom()+" a réussie.",
                        "/Images/checked.png");
            }
         }
@@ -128,36 +165,77 @@ public class EcouteurListeClients implements Initializable {
 
     }
 
-    private void modifierclientSelectionner(Client cl) throws SQLException {
+    private void modifierclientSelectionner(Client cl) throws SQLException, IOException {
         ConnectionClass connectionClass=new ConnectionClass();
         Connection connection=connectionClass.getConnection();
         int fid=cl.isClientFidele()?1:0;
-        String sql="UPDATE client " +
-                "SET nomClient='"+cl.getNomClient()+"', " +
-                "prenomClient='"+cl.getPrenomClient()+"', " +
-                "mailClient='"+cl.getMailClient()+"', " +
-                " clientFidele="+fid+
-                " where idClient='"+cl.getIdClient()+"'";
-        Statement statement = connection.createStatement();
-        int i=statement.executeUpdate(sql);
-
-        statement.execute(sql);
+        String sql="update client set nomClient=? , prenomClient=? , mailClient=? ,clientFidele=?,pdp=? where idClient=?";
+        PreparedStatement statement=connection.prepareStatement(sql);
+        statement.setString(1,cl.getNom());
+        statement.setString(2,cl.getPrenom());
+        statement.setString(3,cl.getMailClient());
+        statement.setInt(4,fid);
+        if(uneImageEstSelectionner){
+            statement.setBinaryStream(5,fis);
+            uneImageEstSelectionner=false;
+        }
+        else{
+            InputStream is=inputClient(cl);
+            statement.setBinaryStream(5,is);
+        }
+        statement.setInt(6,cl.getIdClient());
+        statement.executeUpdate();
         statement.close();
         nettoyageScene();
     }
 
-    public void selectionClient(){
+    private InputStream inputClient(Client cl) throws SQLException {
+        ConnectionClass connectionClass=new ConnectionClass();
+        Connection connection=connectionClass.getConnection();
+        String sqlTEST="SELECT pdp from client where idClient="+cl.getIdClient();
+        Statement stat= connection.createStatement();
+        ResultSet res=stat.executeQuery(sqlTEST);
+        res.next();
+        InputStream is=res.getBinaryStream("pdp");
+        return is;
+    }
+
+    public void selectionClient() throws SQLException, IOException {
         viderChamps();
         if(!table.getSelectionModel().isEmpty()){
             Client clientSelectionner=table.getSelectionModel().getSelectedItem();
-            nom.setText(clientSelectionner.getNomClient());
-            prenom.setText(clientSelectionner.getPrenomClient());
+            selectionImage(clientSelectionner);
+            nom.setText(clientSelectionner.getNom());
+            prenom.setText(clientSelectionner.getPrenom());
             mail.setText(clientSelectionner.getMailClient());
             if(clientSelectionner.isClientFidele()){
                 fidele.setSelected(true);
             }
         }
     }
+    private Image imageClient(Client cl) throws SQLException, IOException {
+        ConnectionClass connectionClass=new ConnectionClass();
+        Connection connection=connectionClass.getConnection();
+        String sql="SELECT pdp from client where idClient="+cl.getIdClient();
+        ResultSet res=connection.createStatement().executeQuery(sql);
+        res.next();
+        InputStream is=res.getBinaryStream("pdp");
+        OutputStream os=new FileOutputStream(new File("photo.jpg"));
+        byte[] content=new byte[1024];
+
+        int size=0;
+        while( (size = is.read(content))!=-1 ){
+            os.write(content,0,size);
+        }
+        os.close();
+        is.close();
+        Image img=new Image("file:photo.jpg",400,300,true, true);
+        return img;
+    }
+    private void selectionImage( Client cl) throws SQLException, IOException {
+        image.setImage(imageClient(cl));
+    }
+
     public void notifBuilder(String titre,String texte,String pathImg){
         Image img=new Image(pathImg);
         Notifications notifBuilder=Notifications.create()
@@ -173,6 +251,7 @@ public class EcouteurListeClients implements Initializable {
         mail.setText("");
         nom.setText("");
         prenom.setText("");
+        image.setImage(new Image("/Images/Photo-non-disponible.jpg"));
         fidele.setSelected(false);
     }
     private Parent root;
@@ -197,8 +276,8 @@ public class EcouteurListeClients implements Initializable {
     @Override
     public void initialize(URL location, ResourceBundle resources) {
         colId.setCellValueFactory(new PropertyValueFactory<Client,Integer>("idClient"));
-        colNom.setCellValueFactory(new PropertyValueFactory<Client,String>("nomClient"));
-        colPrenom.setCellValueFactory(new PropertyValueFactory<Client,String>("prenomClient"));
+        colNom.setCellValueFactory(new PropertyValueFactory<Client,String>("nom"));
+        colPrenom.setCellValueFactory(new PropertyValueFactory<Client,String>("prenom"));
         colMail.setCellValueFactory(new PropertyValueFactory<Client,String>("mailClient"));
         colFidele.setCellValueFactory(new PropertyValueFactory<Client,Boolean>("clientFidele"));
         table.setItems(obList);
